+++ prisma/migrations/20250704201428_added_data_analytics_table/migration.sql
@@ -1,0 +1,26 @@
+-- AlterTable
+ALTER TABLE "ResetPasswordToken" ALTER COLUMN "expDate" SET DEFAULT NOW() + INTERVAL '15 minutes';
+
+-- AlterTable
+ALTER TABLE "ShortUrl" ALTER COLUMN "expDate" SET DEFAULT NOW() + INTERVAL '5 days';
+
+-- CreateTable
+CREATE TABLE "DataAnalytics" (
+    "id" TEXT NOT NULL,
+    "shortUrlId" TEXT NOT NULL,
+    "ownerId" TEXT NOT NULL,
+    "country" TEXT NOT NULL,
+    "city" TEXT NOT NULL,
+    "browser" TEXT NOT NULL,
+    "os" TEXT NOT NULL,
+    "device" TEXT NOT NULL,
+    "accessedAt" TIMESTAMP(3) NOT NULL DEFAULT CURRENT_TIMESTAMP,
+
+    CONSTRAINT "DataAnalytics_pkey" PRIMARY KEY ("id")
+);
+
+-- AddForeignKey
+ALTER TABLE "DataAnalytics" ADD CONSTRAINT "DataAnalytics_shortUrlId_fkey" FOREIGN KEY ("shortUrlId") REFERENCES "ShortUrl"("id") ON DELETE RESTRICT ON UPDATE CASCADE;
+
+-- AddForeignKey
+ALTER TABLE "DataAnalytics" ADD CONSTRAINT "DataAnalytics_ownerId_fkey" FOREIGN KEY ("ownerId") REFERENCES "User"("id") ON DELETE RESTRICT ON UPDATE CASCADE;