+++ src/app/api/redirect/route.ts
@@ -1,9 +1,11 @@
 import { NextResponse } from "next/server";
 import { prisma } from "@/lib/prisma";
 
-export async function GET(req: Request) {
+export async function POST(req: Request) {
   const { searchParams } = new URL(req.url);
   const slug = searchParams.get("slug");
+
+  const body = await req.json();
 
   if (!slug) {
     return NextResponse.json({ error: "Missing slug" }, { status: 400 });
@@ -33,6 +35,33 @@ export async function GET(req: Request) {
       where: { id: urlCheck.id },
       data: { timesUsed: urlCheck.timesUsed + 1 },
     });
+  }
+
+  if(urlCheck.userId) {
+    
+    let device = '';
+    if(["Windows", "Linux", "macOS"].includes(body.os)) {
+      device = "desktop";
+    } else {
+      device = "mobile";
+    } 
+
+    // if the url is owned by a user, record data analytics
+    const dataAnalytics = await prisma.dataAnalytics.create({
+      data: {
+        shortUrlId: urlCheck.id,
+        ownerId: urlCheck.userId,
+        country: body.country || "unknown",
+        city: body.city || "unknown",
+        browser: body.browser || "unknown",
+        os: body.os || "unknown",
+        device:  device || "unknown",
+      }
+    })
+
+    if(!dataAnalytics) {
+      console.log("Failed to record data analytics for slug:", slug);
+    }
   }
 
   const urlToRedirect = urlCheck.originalUrl